breaking: removing old create router declaration in preparation for new backend systemt Signed-off-by: benjdlambert <ben@blam.sh>
