+++ plugins/scaffolder-backend/package.json
@@ -88,7 +88,6 @@
     "@backstage/plugin-scaffolder-node": "workspace:^",
     "@backstage/types": "workspace:^",
     "@opentelemetry/api": "^1.9.0",
-    "@types/express": "^4.17.6",
     "@types/luxon": "^3.0.0",
     "concat-stream": "^2.0.0",
     "express": "^4.17.1",
@@ -122,6 +121,7 @@
     "@backstage/backend-test-utils": "workspace:^",
     "@backstage/cli": "workspace:^",
     "@backstage/plugin-scaffolder-node-test-utils": "workspace:^",
+    "@types/express": "^4.17.6",
     "@types/fs-extra": "^11.0.0",
     "@types/nunjucks": "^3.1.4",
     "@types/supertest": "^2.0.8",
