+++ plugins/scaffolder-backend/report-alpha.api.md
@@ -3,7 +3,6 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { BackendFeature } from '@backstage/backend-plugin-api';
 import { ConditionalPolicyDecision } from '@backstage/plugin-permission-common';
 import { Conditions } from '@backstage/plugin-permission-node';
 import { JsonObject } from '@backstage/types';
@@ -25,10 +24,6 @@ export const createScaffolderTemplateConditionalDecision: (
   permission: ResourcePermission<'scaffolder-template'>,
   conditions: PermissionCriteria<PermissionCondition<'scaffolder-template'>>,
 ) => ConditionalPolicyDecision;
-
-// @alpha (undocumented)
-const _feature: BackendFeature;
-export default _feature;
 
 // @alpha
 export const scaffolderActionConditions: Conditions<{
