+++ plugins/scaffolder-backend/report.api.md
@@ -5,33 +5,23 @@
 ```ts
 import { AuditorService } from '@backstage/backend-plugin-api';
 import { AuthService } from '@backstage/backend-plugin-api';
-import { AutocompleteHandler } from '@backstage/plugin-scaffolder-node/alpha';
 import { BackendFeature } from '@backstage/backend-plugin-api';
 import { BackstageCredentials } from '@backstage/backend-plugin-api';
 import { CatalogApi } from '@backstage/catalog-client';
 import { Config } from '@backstage/config';
-import { CreatedTemplateFilter } from '@backstage/plugin-scaffolder-node/alpha';
-import { CreatedTemplateGlobal } from '@backstage/plugin-scaffolder-node/alpha';
 import { DatabaseService } from '@backstage/backend-plugin-api';
-import { DiscoveryService } from '@backstage/backend-plugin-api';
 import { Duration } from 'luxon';
 import { EventsService } from '@backstage/plugin-events-node';
-import express from 'express';
-import { HttpAuthService } from '@backstage/backend-plugin-api';
 import { HumanDuration } from '@backstage/types';
-import { IdentityApi } from '@backstage/plugin-auth-node';
 import { JsonObject } from '@backstage/types';
 import { JsonValue } from '@backstage/types';
 import { Knex } from 'knex';
-import { LifecycleService } from '@backstage/backend-plugin-api';
 import { Logger } from 'winston';
 import { PermissionEvaluator } from '@backstage/plugin-permission-common';
 import { PermissionRule } from '@backstage/plugin-permission-node';
 import { PermissionRuleParams } from '@backstage/plugin-permission-common';
-import { PermissionsService } from '@backstage/backend-plugin-api';
 import { RESOURCE_TYPE_SCAFFOLDER_ACTION } from '@backstage/plugin-scaffolder-common/alpha';
 import { RESOURCE_TYPE_SCAFFOLDER_TEMPLATE } from '@backstage/plugin-scaffolder-common/alpha';
-import { SchedulerService } from '@backstage/backend-plugin-api';
 import { ScmIntegrations } from '@backstage/integration';
 import { SerializedTask as SerializedTask_2 } from '@backstage/plugin-scaffolder-node';
 import { SerializedTaskEvent as SerializedTaskEvent_2 } from '@backstage/plugin-scaffolder-node';
@@ -248,9 +238,6 @@ export const createFilesystemRenameAction: () => TemplateAction<
   'v1'
 >;
 
-// @public @deprecated
-export function createRouter(options: RouterOptions): Promise<express.Router>;
-
 // @public
 export function createWaitAction(options?: {
   maxWaitTime?: Duration | HumanDuration;
@@ -381,61 +368,6 @@ export type DatabaseTaskStoreOptions = {
   database: DatabaseService | Knex;
   events?: EventsService;
 };
-
-// @public @deprecated
-export interface RouterOptions {
-  // (undocumented)
-  actions?: TemplateAction<any, any, any>[];
-  // (undocumented)
-  additionalTemplateFilters?:
-    | Record<string, TemplateFilter_2>
-    | CreatedTemplateFilter<any, any>[];
-  // (undocumented)
-  additionalTemplateGlobals?:
-    | Record<string, TemplateGlobal_2>
-    | CreatedTemplateGlobal[];
-  // (undocumented)
-  additionalWorkspaceProviders?: Record<string, WorkspaceProvider>;
-  // (undocumented)
-  auditor?: AuditorService;
-  // (undocumented)
-  auth?: AuthService;
-  // (undocumented)
-  autocompleteHandlers?: Record<string, AutocompleteHandler>;
-  // (undocumented)
-  catalogClient: CatalogApi;
-  concurrentTasksLimit?: number;
-  // (undocumented)
-  config: Config;
-  // (undocumented)
-  database: DatabaseService;
-  // (undocumented)
-  discovery?: DiscoveryService;
-  // (undocumented)
-  events?: EventsService;
-  // (undocumented)
-  httpAuth?: HttpAuthService;
-  // (undocumented)
-  identity?: IdentityApi;
-  // (undocumented)
-  lifecycle?: LifecycleService;
-  // (undocumented)
-  logger: Logger;
-  // (undocumented)
-  permissionRules?: Array<
-    TemplatePermissionRuleInput | ActionPermissionRuleInput
-  >;
-  // (undocumented)
-  permissions?: PermissionsService;
-  // (undocumented)
-  reader: UrlReaderService;
-  // (undocumented)
-  scheduler?: SchedulerService;
-  // (undocumented)
-  taskBroker?: TaskBroker_2;
-  // @deprecated (undocumented)
-  taskWorkers?: number;
-}
 
 // @public
 const scaffolderPlugin: BackendFeature;
