+++ plugins/scaffolder-backend/src/ScaffolderPlugin.ts
@@ -134,7 +134,6 @@ export const scaffolderPlugin = createBackendPlugin({
         permissions: coreServices.permissions,
         database: coreServices.database,
         auth: coreServices.auth,
-        discovery: coreServices.discovery,
         httpRouter: coreServices.httpRouter,
         httpAuth: coreServices.httpAuth,
         auditor: coreServices.auditor,
@@ -148,7 +147,6 @@ export const scaffolderPlugin = createBackendPlugin({
         reader,
         database,
         auth,
-        discovery,
         httpRouter,
         httpAuth,
         catalogClient,
@@ -220,7 +218,6 @@ export const scaffolderPlugin = createBackendPlugin({
           additionalTemplateGlobals,
           auth,
           httpAuth,
-          discovery,
           permissions,
           autocompleteHandlers,
           additionalWorkspaceProviders,
