+++ plugins/scaffolder-backend/src/alpha.ts
@@ -13,10 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { scaffolderPlugin } from './ScaffolderPlugin';
-
-/** @alpha */
-const _feature = scaffolderPlugin;
-export default _feature;
-
-export * from './service';
+export * from './service/alpha';
