+++ plugins/scaffolder-backend/src/index.ts
@@ -22,5 +22,8 @@
 
 export { scaffolderPlugin as default } from './ScaffolderPlugin';
 export * from './scaffolder';
-export * from './service/router';
 export * from './lib';
+export {
+  type TemplatePermissionRuleInput,
+  type ActionPermissionRuleInput,
+} from './service/permissions';
